Extracted compiling and execution from listener
