+++ src/main/com/toolbar/listeners/CompileButtonListener.java
@@ -2,16 +2,11 @@ package main.com.toolbar.listeners;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-
 import javax.swing.JTextArea;
 import javax.swing.JTextPane;
 
-import main.com.builder.Libraries;
 import main.com.builder.ScriptBuilder;
-import main.com.ide.MessageHandler;
+import main.com.builder.ScriptCompiler;
 import main.com.ide.Properties;
 
 /**
@@ -27,9 +22,6 @@ public class CompileButtonListener implements ActionListener {
 	// The IDE's notepad and outputs
 	private JTextPane worksheet;
 	private JTextArea buildlog;
-	
-	// For control if compiler fails
-	private boolean buildFailed = false;
 
 	/**
 	 * Class that extracts the text from the worksheet and build
@@ -55,6 +47,7 @@ public class CompileButtonListener implements ActionListener {
 //		String programClasses = "UserScript";
 		
 		ScriptBuilder sb = new ScriptBuilder(properties);
+		ScriptCompiler compiler = new ScriptCompiler(properties, worksheet, buildlog);
 		
 		// Don't compile anything if the script is blank
 		if (worksheet.getText().length() == 0) {
@@ -66,78 +59,8 @@ public class CompileButtonListener implements ActionListener {
 			// Extract code, separate methods/classes, and compile
 			try {
 				sb.buildScript(worksheet);
-				compileScript();
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
+				compiler.compileAndRun();
+			} catch (Exception e) { /* Should not have issues here */ }
 		}
-	}
-
-	/**
-	 *  Calls files necessary to write/compile user classes
-	 * 
-	 * @throws Exception - Any possible error to arise from compiling/execution
-	 */
-	public void compileScript() throws Exception {
-
-		String arguments = properties.getCommandArguments();
-		String OS = getOperatingSystem();
-		String libraries = Libraries.getLibraries(OS);
-		
-		// Compile code
-		buildlog.append("Building script...\r\n");
-		runProcess("javac -cp " + libraries + " .UserFiles/UserScript.java");
-		
-		if (buildFailed) {
-			return;
-		}
-
-		buildlog.append("Complete.\r\n");
-		buildlog.append("---------------------------------------\r\n");
-
-		// Execute.
-		runProcess("java " + arguments + " -cp " + libraries + " UserScript");
-
-		buildlog.append("---------------------------------------\r\n");
-		buildlog.append("Process complete");
-
-	}
-
-	// Generates the process and executes it
-	private void runProcess(String command) throws Exception {
-
-		Process runtimeProcess = Runtime.getRuntime().exec(command);
-		// Print any console outputs in the buildlog
-		MessageHandler.handleProgramMessaage(runtimeProcess.getInputStream(), buildlog);
-
-		// Check for any error messages
-		InputStream errorStream = runtimeProcess.getErrorStream();
-		BufferedReader errorLines = new BufferedReader(new InputStreamReader(errorStream));
-		String errorMessage = errorLines.readLine();
-
-		if (errorMessage != null) {
-
-			while (errorMessage != null) {
-				// If there is an error, print the message and signal build failure
-				MessageHandler.handleErrorMessage(errorMessage, worksheet, buildlog);
-				errorMessage = errorLines.readLine();
-			}
-			
-			buildFailed = true;
-			runtimeProcess.waitFor();
-			return;
-		}
-		// Otherwise command succeed and wait for process to execute.
-		buildFailed = false;
-		runtimeProcess.waitFor();
-	}
-	
-	/**
-	 * Get users operating system
-	 * 
-	 * @return - A string that states the OS.
-	 */
-	private String getOperatingSystem() {
-		return System.getProperty("os.name");
 	}
 }
